refactor: S10.10 fix MISRA 10.4 essential-type mismatches tree-wide
Add U suffix to integer literals (and a handful of anonymous-enum
initializers and named tunable constants) so each operator's operands
share the same MISRA essential-type category. Two byte-comparison sites
in SolidSyslog.c (BOM strip) and SolidSyslogFormatter.c (UTF-8 lead
range) refactored to use `unsigned char` comparisons with hex literals
where the (char) cast did not satisfy cppcheck-misra's essential-type
tracking. SOLIDSYSLOG_FORMATTER_STORAGE_SIZE picks up U on the literal
inside the macro so every expansion is uniform.

cppcheck-misra: rule 10.4 findings 92 → 0.

+++ Core/Source/SolidSyslogFileBlockDevice.c
@@ -12,20 +12,20 @@ struct SolidSyslogFile;
 
 enum
 {
-    MAX_PATH_SIZE = 128
+    MAX_PATH_SIZE = 128U
 };
 
 static const char FILE_EXTENSION[] = ".log";
 
 enum
 {
-    SEQUENCE_DIGITS = 2,
-    FILENAME_SUFFIX = SEQUENCE_DIGITS + sizeof(FILE_EXTENSION) - 1,
-    MAX_PREFIX_LENGTH = MAX_PATH_SIZE - FILENAME_SUFFIX - 1,
+    SEQUENCE_DIGITS = 2U,
+    FILENAME_SUFFIX = SEQUENCE_DIGITS + sizeof(FILE_EXTENSION) - 1U,
+    MAX_PREFIX_LENGTH = (size_t) MAX_PATH_SIZE - (size_t) FILENAME_SUFFIX - 1U,
     /* Two-digit on-disk sequence — indices > 99 cannot be represented
      * uniquely. Without this guard, a wide blockIndex would be narrowed
      * to uint8_t and alias an existing block (256 -> 00). */
-    MAX_BLOCK_INDEX = 99
+    MAX_BLOCK_INDEX = 99U
 };
 
 static inline bool FileBlockDevice_IsValidBlockIndex(size_t blockIndex)
@@ -247,7 +247,7 @@ static inline const char* FileBlockDevice_FormatBlockFilename(
 
     SolidSyslogFormatter_BoundedString(formatter, device->PathPrefix, MAX_PREFIX_LENGTH);
     SolidSyslogFormatter_TwoDigit(formatter, (uint8_t) blockIndex);
-    SolidSyslogFormatter_BoundedString(formatter, FILE_EXTENSION, sizeof(FILE_EXTENSION) - 1);
+    SolidSyslogFormatter_BoundedString(formatter, FILE_EXTENSION, sizeof(FILE_EXTENSION) - 1U);
 
     return SolidSyslogFormatter_AsFormattedBuffer(formatter);
 }
